cheri: fix cast for PR_GET_NAME argument
According to prctl(2), the function is defined as

int prctl(int option, unsigned long arg2,
   unsigned long arg3, unsigned long arg4, unsigned long arg5);

but the manpage clarifies that for PR_SET_NAME, arg2 is cast to (char *).

busybox passes (long)comm as arg2 where comm is a char *. Replace the
cast with uintptr_t to preserve the capability on a cheri system. This
should be safe for non-cheri as well.

+++ libbb/vfork_daemon_rexec.c
@@ -39,7 +39,7 @@ int FAST_FUNC re_execed_comm(void)
 	/* Hopefully (strrchr(e) - e) evaluates to constant at compile time: */
 	expected_comm = bb_busybox_exec_path + (strrchr(e, '/') - e) + 1;
 
-	prctl(PR_GET_NAME, (long)comm, 0, 0, 0);
+	prctl(PR_GET_NAME, (uintptr_t)comm, 0, 0, 0);
 	//bb_error_msg("comm:'%.*s' expected:'%s'", 16, comm, expected_comm);
 	return strcmp(comm, expected_comm) == 0;
 }
@@ -47,7 +47,7 @@ int FAST_FUNC re_execed_comm(void)
 void FAST_FUNC set_task_comm(const char *comm)
 {
 	/* okay if too long (truncates) */
-	prctl(PR_SET_NAME, (long)comm, 0, 0, 0);
+	prctl(PR_SET_NAME, (uintptr_t)comm, 0, 0, 0);
 }
 #endif
